# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OncoSafeRx API (New Clinical/Genomic/AI Endpoints)
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PatientContext:
      type: object
      properties:
        patientId: { type: string }
        age: { type: integer, minimum: 0, maximum: 120 }
        weight: { type: number }
        height: { type: number }
        sex: { type: string, enum: [male, female, other, unknown] }
        comorbidities:
          type: array
          items: { type: string }
        allergies:
          type: array
          items:
            type: object
            properties:
              allergen: { type: string }
              reaction: { type: string }
              severity: { type: string, enum: [mild, moderate, severe] }
        kidneyFunction: { type: number, description: eGFR }
        liverFunction:
          type: object
          properties:
            alt: { type: number }
            ast: { type: number }
            bilirubin: { type: number }
            albumin: { type: number }
        cancerType: { type: string }
        cancerStage: { type: string }
        performanceStatus: { type: integer, minimum: 0, maximum: 4 }
        priorTherapies:
          type: array
          items:
            type: object
            properties:
              therapy: { type: string }
              startDate: { type: string, format: date-time }
              endDate: { type: string, format: date-time }
              response: { type: string }
        biomarkers:
          type: object
          additionalProperties: true
        genetics:
          type: object
          additionalProperties: true
        laboratoryValues:
          type: object
          additionalProperties: true
...
    Variant:
      type: object
      required: [gene]
      properties:
        gene: { type: string }
        alteration: { type: string }
        type: { type: string, enum: [snv, indel, cnv, fusion, structural] }
        impact: { type: string, enum: [high, moderate, low, modifier] }
        clinicalSignificance: { type: string }
        alleleFrequency: { type: number }
    GenomicData:
      type: object
      required: [sampleId]
      properties:
        sampleId: { type: string }
        cancerType: { type: string }
        sampleType: { type: string, enum: [tumor, normal, liquid_biopsy] }
        sequencingPlatform: { type: string }
        panelType: { type: string }
        variants:
          type: array
          items: { $ref: '#/components/schemas/Variant' }
        structuralVariants:
          type: array
          items: { type: object }
        copyNumberVariations:
          type: array
          items: { type: object }
...
paths:
...
  /genomics/profiling/analyze-profile:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [genomicData]
              properties:
                genomicData: { $ref: '#/components/schemas/GenomicData' }
                patientContext: { $ref: '#/components/schemas/PatientContext' }
                analysisOptions: { type: object }
      responses:
        '200': { description: Genomic profile analysis }
  /genomics/profiling/interpret-variants:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variants]
              properties:
                variants:
                  type: array
                  items: { $ref: '#/components/schemas/Variant' }
                cancerType: { type: string }
                includeTherapeuticImplications: { type: boolean }
                includePrognosticImplications: { type: boolean }
                includeResistanceImplications: { type: boolean }
      responses:
        '200': { description: Variant interpretation }
  /genomics/profiling/analyze-tmb:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variants]
              properties:
                variants:
                  type: array
                  items: { $ref: '#/components/schemas/Variant' }
                cancerType: { type: string }
                sequencingInfo: { type: object }
      responses:
        '200': { description: TMB analysis }
  /genomics/profiling/analyze-msi:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                msiMarkers: { type: object }
                variants:
                  type: array
                  items: { $ref: '#/components/schemas/Variant' }
                cancerType: { type: string }
      responses:
        '200': { description: MSI analysis }
  /genomics/profiling/analyze-hrd:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variants]
              properties:
                variants:
                  type: array
                  items: { $ref: '#/components/schemas/Variant' }
                structuralVariants: { type: array, items: { type: object } }
                cancerType: { type: string }
      responses:
        '200': { description: HRD analysis }
  /genomics/profiling/identify-targets:
    post:
      tags: [Genomics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variants:
                  type: array
                  items: { $ref: '#/components/schemas/Variant' }
                cancerType: { type: string }
                evidenceLevels: { type: array, items: { type: string } }
                includeClinicalTrials: { type: boolean }
      responses:
        '200': { description: Therapeutic targets }
...